[Update] Detect i18n message resource by class replace identify message resource by bundle string [Update] Update MessageResource support message update operate [Fix] Logger i18n support didn't work [Release] Release version 1.1.0.1

+++ src/main/java/org/nervousync/security/digest/impl/CRCDigestAdapterImpl.java
@@ -30,7 +30,7 @@ import java.util.*;
  * <h2 class="zh-CN">CRC摘要算法适配器的实现类</h2>
  *
  * @author Steven Wee	<a href="mailto:wmkm0113@Hotmail.com">wmkm0113@Hotmail.com</a>
- * @version $Revision : 1.0 $ $Date: Jan 13, 2012 13:50:21 $
+ * @version $Revision: 1.0.0 $ $Date: Jan 13, 2012 13:50:21 $
  */
 public final class CRCDigestAdapterImpl extends SecureAdapter {
     /**
@@ -223,7 +223,7 @@ public final class CRCDigestAdapterImpl extends SecureAdapter {
      */
     private static long reverseBit(long value, int bit) {
         if (value < 0) {
-            value += Math.pow(2, bit);
+            value += (long) Math.pow(2, bit);
         }
         String reverseValue = new StringBuilder(Long.toString(value, 2)).reverse().toString();
         if (reverseValue.length() < bit) {
